Move converters into separate directory
