+++ cheetah/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqa
-import cheetah.dontbmad
-import cheetah.nocelot
+import cheetah.converters.dontbmad
+import cheetah.converters.nocelot
 from cheetah.accelerator import *
 from cheetah.particles import *
